Added fare
new class Fare and fixed the table
