+++ package.json
@@ -15,6 +15,7 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
+    "react-native-table-component": "^1.2.2",
     "react-native-web": "0.17.7"
   },
   "devDependencies": {
